ping-service/pyproject.toml
[tool.poetry]
name = "ping-thing"
version = "0.1.0"
description = ""
authors = ["Rhys Davies <rhys@memes.nz>"]
readme = "README.md"
packages = [{include = "ping_thing"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
aiohttp = {version = "^3.8.3", extras = ["speedups"]}
asyncstdlib = "^3.10.5"
google-auth = "2.28.2"
requests = "^2.31.0"
boto3 = "^1.34.65"
aws-request-signer = "^1.2.0"
google-cloud-storage = "^2.16.0"
firebase-admin = "^6.5.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^23.1a1", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
